Move cond logic to separate file
